refactor contact information
