+++ client/src/components/fieldsets/shared/ContactInformationFieldset.tsx
@@ -1,60 +1,66 @@
 import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React from 'react'
+import { EmailValidators } from '../../../utils/validators/EmailValidators'
 import { PhoneNumberValidators } from '../../../utils/validators/PhoneNumberValidator'
-import { PostalCodeValidator } from '../../../utils/validators/PostalCodeValidators'
 import Input from '../../Core/DataEntry/Input'
-import RadioButton from '../../Core/DataEntry/RadioButton'
-import StreetNumberAdditionField from '../../Core/DataEntry/StreetNumberAdditionField'
-import Field from '../../Core/Field/Field'
+import ControlField from '../../Core/Field/ControlField'
 import Section from '../../Core/Field/Section'
 import Column from '../../Core/Layout/Column/Column'
-import Row from '../../Core/Layout/Row/Row'
+import { ConnectedFieldsetProps } from '../../hooks/fieldsets/types'
+import { useFieldsetContent } from '../../hooks/fieldsets/useFieldsetContent'
+import { useFieldsetControl } from '../../hooks/fieldsets/useFieldsetControl'
 
-interface Props {
+interface Props extends ConnectedFieldsetProps<Fields> {
     prefillData?: ContactInformationFieldsetModel
     readOnly?: true
 }
 
 export interface ContactInformationFieldsetModel {
-    street: string
-    streetNo: string
-    streetNoAddition: string
-    postalCode: string
-    city: string
-    phoneNumberContactPerson: string
-    contact: string
+    phone: string
+    email: string
 }
+type Fields = 'email' | 'phone'
 
 const ContactInformationFieldset: React.FunctionComponent<Props> = props => {
-    const { prefillData, readOnly } = props
+    const { prefillData, readOnly, fieldNaming, fieldControls } = props
     const { i18n } = useLingui()
+    const content = useFieldsetContent<Fields>(
+        {
+            email: {
+                label: i18n._(t`E-mailadres`),
+                placeholder: i18n._(t`gebruiker@mail.nl`),
+            },
+            phone: {
+                label: i18n._(t`Tel. nr.`),
+                placeholder: i18n._(t`06 - 123 456 78`),
+            },
+        },
+        fieldNaming
+    )
+    const controls = useFieldsetControl<Fields>(
+        {
+            email: {
+                validators: [EmailValidators.isEmailAddress],
+            },
+            phone: {
+                validators: [PhoneNumberValidators.isPhoneNumber],
+            },
+        },
+        fieldControls
+    )
 
     if (readOnly) {
         return (
             <Section title={i18n._(t`Contactgegevens`)}>
                 <Column spacing={4}>
-                    <Field label={i18n._(t`Straatnaam + huisnr.`)} horizontal={true}>
-                        <p>{`${prefillData?.street} ${prefillData?.streetNo} ${
-                            prefillData?.streetNoAddition ? prefillData?.streetNoAddition : ''
-                        }`}</p>
-                    </Field>
+                    <ControlField control={controls.phone} label={content.phone?.label} horizontal={true}>
+                        <p>{prefillData?.phone}</p>
+                    </ControlField>
 
-                    <Field label={i18n._(t`Postcode`)} horizontal={true}>
-                        <p>{prefillData?.postalCode}</p>
-                    </Field>
-
-                    <Field label={i18n._(t`Plaats`)} horizontal={true}>
-                        <p>{prefillData?.city}</p>
-                    </Field>
-
-                    <Field label={i18n._(t`Tel. nr. contactpersoon`)} horizontal={true}>
-                        <p>{prefillData?.phoneNumberContactPerson}</p>
-                    </Field>
-
-                    <Field label={i18n._(t`Contact bij voorkeur`)} horizontal={true}>
-                        <p>{prefillData?.contact}</p>
-                    </Field>
+                    <ControlField control={controls.email} label={content.email?.label} horizontal={true}>
+                        <p>{prefillData?.email}</p>
+                    </ControlField>
                 </Column>
             </Section>
         )
@@ -63,55 +69,23 @@ const ContactInformationFieldset: React.FunctionComponent<Props> = props => {
     return (
         <Section title={i18n._(t`Contactgegevens`)}>
             <Column spacing={4}>
-                <Field label={i18n._(t`Straatnaam + huisnr.`)} horizontal={true}>
-                    <StreetNumberAdditionField
-                        prefillData={{
-                            street: prefillData?.street || '',
-                            streetNr: prefillData?.streetNo || '',
-                            addition: prefillData?.streetNoAddition || '',
-                        }}
-                    />
-                </Field>
-                <Field label={i18n._(t`Postcode`)} horizontal={true}>
+                <ControlField label={content.phone?.label} horizontal={true}>
                     <Input
-                        name="postal-code"
-                        placeholder={i18n._(t`1234 AB`)}
-                        defaultValue={prefillData?.postalCode}
-                        validators={[PostalCodeValidator.isPostalCode]}
+                        name="phone"
+                        placeholder={content.phone?.placeholder}
+                        defaultValue={prefillData?.phone}
+                        validators={controls.phone?.validators}
                     />
-                </Field>
-                <Field label={i18n._(t`Plaats`)} horizontal={true}>
-                    <Input name="city" placeholder={i18n._(t`Plaats`)} defaultValue={prefillData?.city} />
-                </Field>
-                <Field label={i18n._(t`Tel. nr. contactpersoon`)} horizontal={true}>
+                </ControlField>
+
+                <ControlField control={controls.email} label={content.email?.label} horizontal={true}>
                     <Input
-                        name="phone-number"
-                        placeholder={i18n._(t`06 - 123 456 78`)}
-                        defaultValue={prefillData?.phoneNumberContactPerson}
-                        validators={[PhoneNumberValidators.isPhoneNumber]}
+                        name="email"
+                        placeholder={content.email?.placeholder}
+                        defaultValue={prefillData?.email}
+                        validators={controls.email?.validators}
                     />
-                </Field>
-                <Field label={i18n._(t`Contact bij voorkeur`)} horizontal={true}>
-                    <Column spacing={4}>
-                        <Row>
-                            <RadioButton name={'contact-preference'} value="call" />
-                            <p>{i18n._(t`Bellen`)}</p>
-                        </Row>
-                        <Row>
-                            <RadioButton name={'contact-preference'} value="whatsapp" />
-                            <p>{i18n._(t`Whatsapp`)}</p>
-                        </Row>
-                        <Row>
-                            <RadioButton name={'contact-preference'} value="mailen" />
-                            <p>{i18n._(t`Mailen`)}</p>
-                        </Row>
-                        <Row>
-                            <RadioButton name={'contact-preference'} value="mailen" />
-                            <p>{i18n._(t`Anders, namelijk:`)}</p>
-                        </Row>
-                        <Input name="contact-preference-input" placeholder={i18n._(t`Anders`)} />
-                    </Column>
-                </Field>
+                </ControlField>
             </Column>
         </Section>
     )
