+++ client/src/views/Authorized/Management/taalhuis/Data/DataUpdateView.tsx
@@ -16,7 +16,7 @@ import Column from '../../../../../components/Core/Layout/Column/Column'
 import Row from '../../../../../components/Core/Layout/Row/Row'
 import Space from '../../../../../components/Core/Layout/Space/Space'
 import BranchInformationFieldset from '../../../../../components/fieldsets/shared/BranchInformationFieldset'
-import ContactInformationFieldset from '../../../../../components/fieldsets/shared/ContactInformationFieldset'
+import ContactPersonInformationFieldset from '../../../../../components/fieldsets/shared/ContactPersonInformationFieldset'
 import { useMockQuery } from '../../../../../components/hooks/useMockQuery'
 import { useMockMutation } from '../../../../../hooks/UseMockMutation'
 import { routes } from '../../../../../routes/routes'
@@ -101,7 +101,7 @@ const DataUpdateView: React.FunctionComponent<Props> = () => {
                     />
                 </Column>
                 <HorizontalRule />
-                <ContactInformationFieldset
+                <ContactPersonInformationFieldset
                 // prefillData={{
                 //     phoneNumberContactPerson: data?.phoneNumberContactPerson ? data?.phoneNumberContactPerson : '',
                 //     contact: data?.contact ? data.contact : '',
